Added controller-related code for login to AccountController, updated URL path for POST request of login
