+++ src/main/java/com/fdmgroup/backend_streamhub/authenticate/controller/AccountController.java
@@ -1,8 +1,14 @@
 package com.fdmgroup.backend_streamhub.authenticate.controller;
 
+import com.fdmgroup.backend_streamhub.authenticate.dto.LoginRequest;
+import com.fdmgroup.backend_streamhub.authenticate.exceptions.IncorrectPasswordException;
+import com.fdmgroup.backend_streamhub.authenticate.exceptions.IncorrectUsernameOrEmailAddressException;
 import com.fdmgroup.backend_streamhub.authenticate.model.Account;
 import com.fdmgroup.backend_streamhub.authenticate.ApiIndex.ApiResponse;
 import com.fdmgroup.backend_streamhub.authenticate.ApiIndex.ApiResponseAccount;
+import com.fdmgroup.backend_streamhub.authenticate.service.LoginService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,6 +16,11 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 public class AccountController {
+
+    private static final Logger accountControllerLogger = LogManager.getLogger(AccountController.class);
+
+    @Autowired
+    private LoginService loginService;
 
     @GetMapping("/account/testing")
     public ResponseEntity<String> accTesting() {
@@ -46,4 +57,26 @@ public class AccountController {
 
         return response;
     }
+
+    @PostMapping("/account/login/submit")
+    public ResponseEntity<String> registrationAttempt(@RequestBody LoginRequest loginRequest) {
+        try {
+            accountControllerLogger.info("Login attempt | {}", loginRequest.toString());
+            loginService.loginUser(loginRequest);
+            return ResponseEntity.status(HttpStatus.OK).body("Login successful.");
+
+        } catch (IncorrectUsernameOrEmailAddressException e) {
+            accountControllerLogger.error("Unsuccessful login due to incorrect username or email address entered.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid username or email address entered.");
+
+        } catch (IncorrectPasswordException e) {
+            accountControllerLogger.error("Unsuccessful login due to incorrect password entered.");
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid password entered.");
+
+        } catch (Exception e) {
+            accountControllerLogger.fatal("Unsuccessful registration attempt due to invalid an unexpected error.");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Unexpected error occurred.");
+        }
+    }
+
 }
